+++ Modules/Intent.Modules.AspNetCore.OData.EntityFramework/NugetPackages.cs
@@ -23,7 +23,7 @@ namespace Intent.Modules.AspNetCore.OData.EntityFramework
                             .WithNugetDependency("Microsoft.OData.Edm", "8.2.3")
                             .WithNugetDependency("Microsoft.OData.ModelBuilder", "2.0.0")
                             .WithNugetDependency("Microsoft.Spatial", "8.2.3"),
-                        ( >= 6, >= 0) => new PackageVersion("8.3.0")
+                        ( >= 6, >= 0) => new PackageVersion("8.3.1")
                             .WithNugetDependency("Microsoft.OData.Core", "7.21.6")
                             .WithNugetDependency("Microsoft.OData.Edm", "7.21.6")
                             .WithNugetDependency("Microsoft.OData.ModelBuilder", "1.0.9")