+++ Modules/Intent.Modules.AzureFunctions/NugetPackages.cs
@@ -141,11 +141,11 @@ namespace Intent.Modules.AzureFunctions
             NugetRegistry.Register(MicrosoftAzureFunctionsWorkerExtensionsEventHubsPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 2, >= 0) => new PackageVersion("6.4.0")
+                        ( >= 2, >= 0) => new PackageVersion("6.5.0")
                             .WithNugetDependency("Azure.Messaging.EventHubs", "5.11.5")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Core", "1.19.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Extensions.Abstractions", "1.3.0")
-                            .WithNugetDependency("Microsoft.Extensions.Azure", "1.7.5"),
+                            .WithNugetDependency("Microsoft.Extensions.Azure", "1.12.0"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAzureFunctionsWorkerExtensionsEventHubsPackageName}'"),
                     }
                 );
@@ -185,13 +185,13 @@ namespace Intent.Modules.AzureFunctions
             NugetRegistry.Register(MicrosoftAzureFunctionsWorkerExtensionsServiceBusPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 2, >= 0) => new PackageVersion("5.22.2")
-                            .WithNugetDependency("Azure.Identity", "1.12.0")
+                        ( >= 2, >= 0) => new PackageVersion("5.23.0")
+                            .WithNugetDependency("Azure.Identity", "1.13.1")
                             .WithNugetDependency("Azure.Messaging.ServiceBus", "7.18.1")
                             .WithNugetDependency("Google.Protobuf", "3.27.1")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Extensions.Abstractions", "1.3.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Extensions.Rpc", "1.0.1")
-                            .WithNugetDependency("Microsoft.Extensions.Azure", "1.7.5"),
+                            .WithNugetDependency("Microsoft.Extensions.Azure", "1.12.0"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAzureFunctionsWorkerExtensionsServiceBusPackageName}'"),
                     }
                 );
@@ -217,7 +217,7 @@ namespace Intent.Modules.AzureFunctions
             NugetRegistry.Register(MicrosoftAzureFunctionsWorkerSdkPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 2, >= 0) => new PackageVersion("2.0.4")
+                        ( >= 2, >= 0) => new PackageVersion("2.0.5")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Sdk.Analyzers", "1.2.2")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Sdk.Generators", "1.3.5"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAzureFunctionsWorkerSdkPackageName}'"),