+++ Modules/Intent.Modules.ModularMonolith.Host/Templates/ModuleInstallerInterface/ModuleInstallerInterfaceTemplatePartial.cs
@@ -1,12 +1,15 @@
-using System;
-using System.Collections.Generic;
 using Intent.Engine;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
+using Intent.Modules.Common.CSharp.Nuget;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.Common.VisualStudio;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
@@ -21,10 +24,12 @@ namespace Intent.Modules.ModularMonolith.Host.Templates.ModuleInstallerInterface
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         public ModuleInstallerInterfaceTemplate(IOutputTarget outputTarget, object model = null) : base(TemplateId, outputTarget, model)
         {
-            AddNugetDependency(NugetPackages.SwashbuckleAspNetCore(outputTarget));
-            AddNugetDependency(NugetPackages.MassTransit(outputTarget));
-
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
+                .OnBuild(c =>
+                {
+                    AddNugetDependency(NugetPackages.SwashbuckleAspNetCore(outputTarget));
+                    AddNugetDependency(NugetPackages.MassTransit(outputTarget));
+                })
                 .AddInterface($"IModuleInstaller", @interface =>
                 {
                     CSharpFile.AddUsing("Microsoft.Extensions.Configuration");
@@ -45,7 +50,7 @@ namespace Intent.Modules.ModularMonolith.Host.Templates.ModuleInstallerInterface
                         method.AddParameter("IRegistrationConfigurator", "cfg");
                     });
                 })
-                .AddClass("IModuleInstallerExtensions", @class => 
+                .AddClass("IModuleInstallerExtensions", @class =>
                 {
                     @class.Static();
                     @class.AddMethod("void", "ConfigureContainer", method =>
@@ -80,6 +85,26 @@ namespace Intent.Modules.ModularMonolith.Host.Templates.ModuleInstallerInterface
                     });
 
                 });
+        }
+
+        public override void AfterTemplateRegistration()
+        {
+            base.AfterTemplateRegistration();
+
+            // This block is because version 8.5.0 of MassTransit is only compatible with EF9 (not EF8). 
+            // With Modular Monolith, EF is installed in the Modules, so we cannot know if EF is being installed nor which version.
+            // We are making the assumption that if the Framework is NET8, then EF8 will be used and thus MT 8.4.1 must be used
+            if (!OutputTarget.Parent.TargetFramework().Contains("9.0"))
+            { 
+                NugetRegistry.Register("MassTransit.Abstractions", v => new PackageVersion("8.4.1", true));
+                NugetRegistry.Register(NugetPackages.MassTransitPackageName, v => new PackageVersion("8.4.1", true)
+                        .WithNugetDependency("MassTransit.Abstractions", "8.4.1")
+                        .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "8.0.0")
+                        .WithNugetDependency("Microsoft.Extensions.Diagnostics.HealthChecks", "8.0.0")
+                        .WithNugetDependency("Microsoft.Extensions.Hosting.Abstractions", "8.0.0")
+                        .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "8.0.0")
+                        .WithNugetDependency("Microsoft.Extensions.Options", "8.0.0"));
+            }
         }
 
         [IntentManaged(Mode.Fully)]