+++ Modules/Intent.Modules.AspNetCore.IntegrationTesting/release-notes.md
@@ -1,6 +1,8 @@
 ### Version 2.0.4
 
+
 - Improvement: Adjusted text namespaces to align with folders, which were causing xunit issues in some scenarios.
+- Improvement: Updated NuGet package versions.
 
 ### Version 2.0.3