+++ Modules/Intent.Modules.HotChocolate.GraphQL.AspNetCore/release-notes.md
@@ -1,3 +1,7 @@
+### Version 6.0.4
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 6.0.3
 
 - Improvement: Updated NuGet package versions.